requirements.txt
Django==1.3.1
distribute==0.6.24
django-annoying==0.7.6
django-grappelli==2.3.7
django-registration==0.7
django-registration-defaults==0.3
psycopg2==2.4.4
wsgiref==0.1.2
